requirements.txt
fastapi
transformers
torch
uvicorn
pillow